feat: Add the Category Controller and adjust the Service methods
Closes #13

+++ src/main/java/com/github/arthurscarpin/balancee/domain/category/service/CategoryService.java
@@ -50,6 +50,11 @@ public class CategoryService {
 
     @Transactional
     public void deleteById(Long id) {
-        repository.deleteById(id);
+        Optional<Category> categoryExists = repository.findById(id);
+        if (categoryExists.isPresent()) {
+            repository.deleteById(id);
+        } else {
+            throw new BusinessException("Category not found!");
+        }
     }
 }
